# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API server for Ondaum",
        "title": "Ondaum API",
        "contact": {},
        "version": "1.0"
    },
    "host": "ondaum.revimal.me",
    "basePath": "/api/v1",
    "paths": {
...
        "/chat/:session_id": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get chat with histories and summary",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get chat",
                "operationId": "GetChat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chat.ChatWithSummaryAndHistoriesDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    }
                }
            }
        },
        "/chat/:session_id/archive": {
            "post": {
                "description": "Archive a chat to prevent it from being accessed again and allow to summarize it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Archive a chat",
                "operationId": "ArchiveChat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chat.ArchiveChatHandlerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    }
                }
            }
        },
...
        "/chats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List chats with summaries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "List chats",
                "operationId": "ListChat",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/chat.ChatWithSummaryDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    }
                }
            }
        },
...
        "/user/privacy": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates or creates the user's privacy information including gender and birthday.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update or Create User Privacy Information",
                "operationId": "UpsertUserPrivacy",
                "parameters": [
                    {
                        "description": "User privacy information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpsertPrivacyHandlerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UpsertPrivacyHandlerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    }
                }
            }
        },
        "/user/self": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's profile, including basic information and optional onboarding data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get Current Authenticated User Profile",
                "operationId": "GetSelfUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.SimplifiedUserDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "chat.ArchiveChatHandlerResponse": {
            "type": "object",
            "properties": {
                "finished": {
                    "type": "boolean"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "chat.ChatWithSummaryAndHistoriesDTO": {
            "type": "object",
            "properties": {
                "histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/chat.HistoryDTO"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_archived": {
                    "type": "boolean"
                },
                "is_finished": {
                    "type": "boolean"
                },
                "session_id": {
                    "type": "string"
                },
                "started_date": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/chat.SummaryDTO"
                },
                "user_id": {
                    "type": "string"
                },
                "user_timezone": {
                    "type": "string"
                }
            }
        },
        "chat.ChatWithSummaryDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_archived": {
                    "type": "boolean"
                },
                "is_finished": {
                    "type": "boolean"
                },
                "session_id": {
                    "type": "string"
                },
                "started_date": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/chat.SummaryDTO"
                },
                "user_id": {
                    "type": "string"
                },
                "user_timezone": {
                    "type": "string"
                }
            }
        },
        "chat.HistoryDTO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "role": {
                    "type": "string"
                },
                "when": {
                    "type": "string"
                }
            }
        },
        "chat.SummaryDTO": {
            "type": "object",
            "properties": {
                "emotions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.EmotionRate"
                    }
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "common.Emotion": {
            "type": "string",
            "enum": [
                "joy",
                "sadness",
                "anger",
                "surprise",
                "fear",
                "disgust",
                "neutral"
            ],
            "x-enum-varnames": [
                "EmotionJoy",
                "EmotionSadness",
                "EmotionAnger",
                "EmotionSurprise",
                "EmotionFear",
                "EmotionDisgust",
                "EmotionNeutral"
            ]
        },
        "common.EmotionRate": {
            "type": "object",
            "properties": {
                "emotion": {
                    "$ref": "#/definitions/common.Emotion"
                },
                "rate": {
                    "type": "number"
                }
            }
        },
...
        "http.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "user.SimplifiedAdditionDTO": {
            "type": "object",
            "properties": {
                "concerns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "emotions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.Emotion"
                    }
                }
            }
        },
        "user.SimplifiedPrivacyDTO": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                }
            }
        },
        "user.SimplifiedUserDTO": {
            "type": "object",
            "properties": {
                "addition": {
                    "$ref": "#/definitions/user.SimplifiedAdditionDTO"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "privacy": {
                    "$ref": "#/definitions/user.SimplifiedPrivacyDTO"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "user.UpsertPrivacyHandlerRequest": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                }
            }
        },
        "user.UpsertPrivacyHandlerResponse": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "boolean"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}